--- lib/epoch_generator.go
@@ -1,7 +1,0 @@
-package lib
-
-import "context"
-
-type EpochGenerator interface {
-	NewEpoch(ctx context.Context) (epoch, error)
-}
