+++ lib/master_test.go
@@ -1,0 +1,2 @@
+package lib
+
